Moved operands to package
