+++ src/main/java/x64/instructions/LoadEffectiveAddressInstruction.java
@@ -1,7 +1,7 @@
 package x64.instructions;
 
-import x64.DestinationOperand;
-import x64.SourceOperand;
+import x64.operands.DestinationOperand;
+import x64.operands.SourceOperand;
 
 public class LoadEffectiveAddressInstruction extends BinaryInstruction {
